move models from api class
