+++ src/main/java/me/joshmcfarlin/cryptocompareapi/api/News.java
@@ -4,6 +4,7 @@ import com.google.gson.Gson;
 import com.google.gson.annotations.SerializedName;
 import com.google.gson.reflect.TypeToken;
 import me.joshmcfarlin.cryptocompareapi.CryptoCompareAPIConstant;
+import me.joshmcfarlin.cryptocompareapi.models.news.NewsProvider;
 import me.joshmcfarlin.cryptocompareapi.models.news.NewsStory;
 import me.joshmcfarlin.cryptocompareapi.utils.CallTypes;
 import me.joshmcfarlin.cryptocompareapi.utils.Connection;
@@ -42,63 +43,5 @@ public class News {
 
         Type newsListType = new TypeToken<List<NewsStory>>() {}.getType();
         return new Gson().fromJson(r, newsListType);
-    }
-
-    /**
-     * Represents a news provider on CryptoCompare's website
-     */
-    public class NewsProvider {
-        /**
-         * Key name used by CryptoCompare
-         */
-        private String key;
-
-        /**
-         * Name provided on CryptoCompare's website
-         */
-        private String name;
-
-        /**
-         * Language used by news provided
-         */
-        private String lang;
-
-        /**
-         * Url to image representing news provider
-         */
-        private String img;
-
-        @Override
-        public String toString() {
-            return String.format("Key: %s, Name: %s, Lang: %s", key, name, lang);
-        }
-
-        /**
-         * {@link NewsProvider#key}
-         */
-        public String getKey() {
-            return key;
-        }
-
-        /**
-         * {@link NewsProvider#name}
-         */
-        public String getName() {
-            return name;
-        }
-
-        /**
-         * {@link NewsProvider#lang}
-         */
-        public String getLang() {
-            return lang;
-        }
-
-        /**
-         * {@link NewsProvider#img}
-         */
-        public String getImg() {
-            return img;
-        }
     }
 }
